TOBAGO-2060: Tobago should be compatible with Spring Boot
* Replacing CDI.current()
* Refactoring TobagoConfig and its initialization
