+++ tobago-core/src/test/java/org/apache/myfaces/tobago/internal/config/TobagoConfigMergingUnitTest.java
@@ -19,57 +19,45 @@
 
 package org.apache.myfaces.tobago.internal.config;
 
+import org.apache.myfaces.tobago.config.TobagoConfig;
 import org.apache.myfaces.tobago.context.Theme;
 import org.apache.myfaces.tobago.context.ThemeScript;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.xml.sax.SAXException;
 
-import javax.xml.parsers.ParserConfigurationException;
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 public class TobagoConfigMergingUnitTest {
 
   @Test
-  public void testPreventFrameAttacksCascadingDefault()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testPreventFrameAttacksCascadingDefault() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml",
-        "tobago-config-merge-1.xml");
+    final TobagoConfig config =
+        new TobagoConfig(null, "tobago-config-merge-0.xml", "tobago-config-merge-1.xml");
 
     Assertions.assertFalse(config.isPreventFrameAttacks());
   }
 
   @Test
-  public void testPreventFrameAttacks()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testPreventFrameAttacks() {
 
-    final TobagoConfigImpl config = load("tobago-config-merge-0.xml");
+    final TobagoConfig config = new TobagoConfig(null, "tobago-config-merge-0.xml");
 
     Assertions.assertFalse(config.isPreventFrameAttacks());
   }
 
   @Test
-  public void testPreventFrameAttacksDefault()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testPreventFrameAttacksDefault() {
 
-    final TobagoConfigImpl config = load("tobago-config-merge-1.xml");
+    final TobagoConfig config = new TobagoConfig(null, "tobago-config-merge-1.xml");
 
     Assertions.assertTrue(config.isPreventFrameAttacks());
   }
 
   @Test
-  public void testContentSecurityPolicy()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testContentSecurityPolicy() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml");
+    final TobagoConfig config = new TobagoConfig(null, "tobago-config-merge-0.xml");
 
     Assertions.assertSame(config.getContentSecurityPolicy().getMode(), ContentSecurityPolicy.Mode.ON);
     final Map<String, String> directiveMap = config.getContentSecurityPolicy().getDirectiveMap();
@@ -78,12 +66,10 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testContentSecurityPolicyExtend()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testContentSecurityPolicyExtend() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml",
-        "tobago-config-merge-1.xml");
+    final TobagoConfig config =
+        new TobagoConfig(null, "tobago-config-merge-0.xml", "tobago-config-merge-1.xml");
 
     Assertions.assertSame(config.getContentSecurityPolicy().getMode(), ContentSecurityPolicy.Mode.REPORT_ONLY);
     final Map<String, String> directiveMap = config.getContentSecurityPolicy().getDirectiveMap();
@@ -93,12 +79,9 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testContentSecurityPolicyOff()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testContentSecurityPolicyOff() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml",
-        "tobago-config-merge-1.xml",
+    final TobagoConfig config = new TobagoConfig(null, "tobago-config-merge-0.xml", "tobago-config-merge-1.xml",
         "tobago-config-merge-2.xml");
 
     Assertions.assertSame(config.getContentSecurityPolicy().getMode(), ContentSecurityPolicy.Mode.OFF);
@@ -106,12 +89,10 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testContentSecurityPolicyNameAttribute()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testContentSecurityPolicyNameAttribute() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml",
-        "tobago-config-merge-3.xml");
+    final TobagoConfig config =
+        new TobagoConfig(null, "tobago-config-merge-0.xml", "tobago-config-merge-3.xml");
 
     Assertions.assertSame(config.getContentSecurityPolicy().getMode(), ContentSecurityPolicy.Mode.ON);
     final Map<String, String> directiveMap = config.getContentSecurityPolicy().getDirectiveMap();
@@ -120,12 +101,9 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testMimeTypes()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testMimeTypes() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-merge-0.xml",
-        "tobago-config-merge-1.xml",
+    final TobagoConfig config = new TobagoConfig(null, "tobago-config-merge-0.xml", "tobago-config-merge-1.xml",
         "tobago-config-merge-2.xml");
 
     final Map<String, String> mimeTypes = config.getMimeTypes();
@@ -136,11 +114,10 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testResourcePriority()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testResourcePriority() {
 
-    final TobagoConfigImpl config = load(
-        "tobago-config-5.0.xml", "tobago-config-5.0-replace.xml");
+    final TobagoConfig config =
+        new TobagoConfig(null, "tobago-config-5.0.xml", "tobago-config-5.0-replace.xml");
 
     final String[] expected = new String[]{
         "script-1.js",
@@ -171,12 +148,10 @@ public class TobagoConfigMergingUnitTest {
   }
 
   @Test
-  public void testMergeThemePatch()
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
+  public void testMergeThemePatch() {
 
-    final TobagoConfigImpl config12 = load(
-        "tobago-config-theme-merge-1.xml",
-        "tobago-config-theme-merge-2.xml");
+    final TobagoConfig config12 =
+        new TobagoConfig(null, "tobago-config-theme-merge-1.xml", "tobago-config-theme-merge-2.xml");
     Assertions.assertEquals("2.0", config12.getDefaultTheme().getVersion());
 
     final ThemeScript[] scripts12 = config12.getDefaultTheme().getScriptResources(false);
@@ -184,9 +159,8 @@ public class TobagoConfigMergingUnitTest {
     Assertions.assertEquals("script-1", scripts12[0].getName());
     Assertions.assertEquals("script-2", scripts12[1].getName());
 
-    final TobagoConfigImpl config21 = load(
-        "tobago-config-theme-merge-2.xml",
-        "tobago-config-theme-merge-1.xml");
+    final TobagoConfig config21 =
+        new TobagoConfig(null, "tobago-config-theme-merge-2.xml", "tobago-config-theme-merge-1.xml");
     Assertions.assertEquals("2.0", config21.getDefaultTheme().getVersion());
 
     final ThemeScript[] scripts21 = config21.getDefaultTheme().getScriptResources(false);
@@ -195,22 +169,4 @@ public class TobagoConfigMergingUnitTest {
     Assertions.assertEquals("script-2", scripts21[1].getName());
   }
 
-  public static TobagoConfigImpl load(final String... names)
-      throws IOException, SAXException, ParserConfigurationException, URISyntaxException {
-
-    final List<TobagoConfigFragment> list = new ArrayList<>();
-
-    for (final String name : names) {
-      final URL url = TobagoConfigMergingUnitTest.class.getClassLoader().getResource(name);
-      final TobagoConfigParser parser = new TobagoConfigParser();
-      list.add(parser.parse(url));
-    }
-
-    final TobagoConfigSorter sorter = new TobagoConfigSorter(list);
-    final TobagoConfigMerger merger = new TobagoConfigMerger(sorter.topologicalSort());
-    final TobagoConfigImpl result = merger.merge();
-    result.resolveThemes();
-    result.lock();
-    return result;
-  }
 }
